+++ CHANGES.md
@@ -10,6 +10,9 @@ in a similar manner as the "regular" data. Please read the relevant section in t
 * The user can use a registered `SessionDisconnectHandler` in order be informed and also intervene in cases
 where the code decides to disconnect the session due to various protocol or configuration parameters violations.
 
+* `ScpFileOpener#getMatchingFilesToSend` has been modified to accept a `Path` as the base directory
+and also return an `Iterable<Path>`.
+
 ## Behavioral changes and enhancements
 
 * [SSHD-882](https://issues.apache.org/jira/browse/SSHD-882) - Provide hooks to allow users to register a consumer
@@ -17,3 +20,5 @@ for STDERR data sent via the `ChannelSession` - especially for the SFTP subsyste
 
 * [SSHD=892](https://issues.apache.org/jira/browse/SSHD-882) - Inform user about possible session disconnect prior
 to disconnecting and allow intervention via `SessionDisconnectHandler`.
+
+* [SSHD-893] Using Path(s) instead of String(s) as DirectoryScanner results